refactor: split rollback.rs / ple/layer.rs under the 500 LoC cap
rollback.rs (511) -> rollback_rewind.rs child: rewind_buffers +
apply_rollback (pure token-buffer rewind + ActiveSeq glue), re-exported.
ple/layer.rs (501) -> seq_state.rs child: PleSeqState with pub(super)
fields, re-exported. No behaviour change.
